# Improve HTTP/2 headers enumerator performance