Optimize enumerating over HTTP/2 headers. Used when writing headers/trailers to HTTP/2 frames.

Before
```
|                              Method |      Mean |    Error |   StdDev |         Op/s | Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------------------------ |----------:|---------:|---------:|-------------:|------:|------:|------:|----------:|
|     KnownSingleValueResponseHeaders |  77.41 ns | 1.154 ns | 1.080 ns | 12,917,659.5 |     - |     - |     - |         - |
|   KnownMultipleValueResponseHeaders |  77.19 ns | 1.182 ns | 1.106 ns | 12,954,674.9 |     - |     - |     - |         - |
|   UnknownSingleValueResponseHeaders | 112.03 ns | 1.253 ns | 1.046 ns |  8,925,975.1 |     - |     - |     - |         - |
| UnknownMultipleValueResponseHeaders | 114.49 ns | 1.598 ns | 1.495 ns |  8,734,013.2 |     - |     - |     - |         - |
```

After
```
|                              Method |     Mean |    Error |   StdDev |         Op/s | Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------------------------ |---------:|---------:|---------:|-------------:|------:|------:|------:|----------:|
|     KnownSingleValueResponseHeaders | 53.02 ns | 0.763 ns | 0.677 ns | 18,860,872.2 |     - |     - |     - |         - |
|   KnownMultipleValueResponseHeaders | 59.02 ns | 1.220 ns | 1.141 ns | 16,942,032.6 |     - |     - |     - |         - |
|   UnknownSingleValueResponseHeaders | 86.60 ns | 1.768 ns | 2.104 ns | 11,547,743.5 |     - |     - |     - |         - |
| UnknownMultipleValueResponseHeaders | 95.04 ns | 1.245 ns | 1.104 ns | 10,521,379.5 |     - |     - |     - |         - |
```